refactor: move sessionstorage utils to sessionstorage dir
