+++ src/pages/[eventAbbr]/ui/get-point.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router'
 import { useInitSetup } from '../../../components/hooks/useInitSetup'
 import { useEffect } from 'react'
 import { Layout } from '../../../components/Layout'
-import { setTrailMapOpenNext } from '../../../util/trailMap'
+import { setTrailMapOpenNext } from '../../../util/sessionstorage/trailMap'
 import { usePostApiV1ProfileByProfileIdPointMutation } from '../../../generated/dreamkast-api.generated'
 import { settingsSelector } from '../../../store/settings'
 import { useSelector } from 'react-redux'
